feat: Refactor user DAO dependency and implementation
* Moved user DAO functionality to a new user_dao.py file for better organization.
* Updated import statement in the init file to reflect the new user DAO location.
* Enhanced the get_user_dao function to provide a request-scoped UserDAO instance using FastAPI's dependency injection.
