+++ backend/src/expenses_counter/modules/middlewares/dependencies/daos/user_dao.py
